UPSTREAM: <carry>: extend termination events
- we tie the shutdown events with the UID of  the first
  (shutdown initiated), this provides us with a more
  deterministic way to compute shutdown duration from
  these events

- move code snippets from the upstream file to openshift
  specific patch file, it reduces chance of code conflict

Note for rebase: squash it into the following commit
cfbb6d6711a UPSTREAM: <carry>: create termination events
