+++ staging/src/k8s.io/apiserver/pkg/server/config.go
@@ -324,11 +324,6 @@ type Config struct {
 	SendRetryAfterWhileNotReadyOnce bool
 }
 
-// EventSink allows to create events.
-type EventSink interface {
-	Create(event *corev1.Event) (*corev1.Event, error)
-}
-
 type RecommendedConfig struct {
 	Config
 
@@ -828,7 +823,9 @@ func (c completedConfig) New(name string, delegationTarget DelegationTarget) (*G
 
 		muxAndDiscoveryCompleteSignals: map[string]<-chan struct{}{},
 
-		eventSink: c.EventSink,
+		OpenShiftGenericAPIServerPatch: OpenShiftGenericAPIServerPatch{
+			eventSink: c.EventSink,
+		},
 	}
 
 	ref, err := eventReference()
